move devtools types out of sinks
