--- libs/rx-store/src/lib/rx-store.ts
@@ -1,4 +1,0 @@
-export * from './store-value';
-export * from './effect';
-export * from './sources';
-export * from './sinks';
