+++ libs/rx-store/src/lib/sinks.ts
@@ -7,27 +7,7 @@ import {
   Subject,
   ReplaySubject,
 } from 'rxjs';
-
-declare global {
-  interface Window {
-    __rxStoreLinks: Subject<{
-      from: { type: 'subject' | 'effect'; name: string };
-      to: { type: 'subject' | 'effect'; name: string };
-    }>;
-    __rxStoreEffects: Subject<{
-      name: string;
-      event: 'spawn' | 'teardown';
-    }>;
-    __rxStoreSubjects: Subject<{
-      name: string;
-    }>;
-    __rxStoreValues: Subject<{
-      from: { type: 'subject' | 'effect'; name: string };
-      to: { type: 'subject' | 'effect'; name: string };
-      value: any;
-    }>;
-  }
-}
+import { ensureDevtools } from './devtools';
 
 /**
  * Sinks are a write-only interface into the subjects
@@ -91,18 +71,3 @@ export const createSinks = <StoreValue extends {}>(
     {}
   ) as Sinks<StoreValue>;
 };
-
-export function ensureDevtools() {
-  if (!window.__rxStoreValues) {
-    window.__rxStoreValues = new Subject<any>();
-  }
-  if (!window.__rxStoreLinks) {
-    window.__rxStoreLinks = new ReplaySubject<any>();
-  }
-  if (!window.__rxStoreEffects) {
-    window.__rxStoreEffects = new ReplaySubject<any>();
-  }
-  if (!window.__rxStoreSubjects) {
-    window.__rxStoreSubjects = new ReplaySubject<any>();
-  }
-}
